requirements.txt
mysql-connector-python==8.1.0
protobuf==4.21.12
openpyxl==3.0.10
pandas==1.3.5
PySimpleGUI==4.60.1
